+++ src/main/java/com/devcourse/web2_1_dashbunny_be/feature/order/controller/dto/DeclineOrdersResponseDto.java
@@ -1,4 +1,4 @@
-package com.devcourse.web2_1_dashbunny_be.feature.user.dto.order.controller.dto;
+package com.devcourse.web2_1_dashbunny_be.feature.order.controller.dto;
 
 import com.devcourse.web2_1_dashbunny_be.domain.user.Orders;
 import com.devcourse.web2_1_dashbunny_be.domain.user.role.DeclineReasonType;
