+++ src/main/java/com/devcourse/web2_1_dashbunny_be/feature/order/repository/OrdersRepository.java
@@ -1,4 +1,4 @@
-package com.devcourse.web2_1_dashbunny_be.feature.user.dto.order.repository;
+package com.devcourse.web2_1_dashbunny_be.feature.order.repository;
 
 import com.devcourse.web2_1_dashbunny_be.domain.user.Orders;
 import org.springframework.data.jpa.repository.JpaRepository;
